quad_test.go
package engine

import (
	"github.com/ghthor/gospec/src/gospec"
	. "github.com/ghthor/gospec/src/gospec"
)

func DescribeAABB(c gospec.Context) {
	aabb := AABB{
		WorldCoord{0, 0},
		WorldCoord{0, 0},
	}

	c.Specify("the width, height and area of an aabb", func() {
		c.Expect(aabb.Width(), Equals, 1)
		c.Expect(aabb.Height(), Equals, 1)
		c.Expect(aabb.Area(), Equals, 1)

		aabb = AABB{
			WorldCoord{0, 0},
			WorldCoord{1, -1},
		}
		c.Expect(aabb.Width(), Equals, 2)
		c.Expect(aabb.Height(), Equals, 2)
		c.Expect(aabb.Area(), Equals, 4)
	})

	c.Specify("aabb contains a coord inside of itself", func() {
		c.Expect(aabb.Contains(WorldCoord{0, 0}), IsTrue)
		containsCheck := func(aabb AABB) {
			for i := aabb.TopL.X; i <= aabb.BotR.X; i++ {
				for j := aabb.TopL.Y; j >= aabb.BotR.Y; j-- {
					c.Expect(aabb.Contains(WorldCoord{i, j}), IsTrue)
				}
			}
		}

		containsCheck(AABB{
			WorldCoord{0, 0},
			WorldCoord{1, -1},
		})
		containsCheck(AABB{
			WorldCoord{1, 1},
			WorldCoord{2, -10},
		})
	})

	c.Specify("flip topleft and bottomright if they are inverted", func() {
		aabb = AABB{
			WorldCoord{0, 0},
			WorldCoord{-1, 1},
		}

		c.Expect(aabb.IsInverted(), IsTrue)
		c.Expect(aabb.Invert().IsInverted(), IsFalse)
	})
}
